intvar/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic[email]==2.8.2
